Implemented remote controllers
