+++ meNexus/web/src/api/hooks/useCreateRemotePost.js
@@ -1,16 +1,16 @@
 import useAxios from './useAxios.js';
 import {ENDPOINTS} from '../config.js'
 
-const useCreateSynapsePost = (refreshPosts) => {
-    console.log("useCreateSynapsePost called");
+const useCreateRemotePost = (refreshPosts) => {
+    console.log("useCreateRemotePost called");
     const { sendRequest, loading, error } = useAxios();
 
-    const createSynapsePost = async (post) => {
+    const createRemotePost = async (post) => {
         console.log('createSynapsePost called for post:', post);
         try {
             const response = await sendRequest({
                 method: 'POST',
-                url: ENDPOINTS.CREATE_SYNAPSE_POST,
+                url: ENDPOINTS.CREATE_REMOTE_POST,
                 data: post
             });
             console.log("Post created:", response);
@@ -23,10 +23,10 @@ const useCreateSynapsePost = (refreshPosts) => {
     };
 
     return {
-        createSynapsePost,
+        createRemotePost,
         loading,
         error
     };
 };
 
-export default useCreateSynapsePost;
+export default useCreateRemotePost;
