+++ synapse/package.json
@@ -16,6 +16,7 @@
     "#protocols/*": "./src/protocols/*",
     "#src/*": "./src/*",
     "#handlers/*": "./src/handlers/*",
+    "#remotes/*": "./src/remotes/*",
     "#utils/*": "./src/utils/*",
     "#uploads/*": "./uploads/*"
   },
