+++ synapse/src/handlers/handleData.js
@@ -15,7 +15,7 @@ export const handleData = async (libp2p, message) => {
                     console.log(`Received SYNAPSE_METADATA request from ${message.meta.sender}.`);
                     const response = await sendRequest({
                         method: 'GET',
-                        url: ENDPOINTS.GET_LOCAL_SYNAPSE_METADATA,
+                        url: ENDPOINTS.GET_SYNAPSE_METADATA,
                         withCredentials: true,
                     });
                     console.log("GET_SYNAPSE_METADATA response ", response);
